has-unique-proxy: Narrow down meaning
All the client needs is that the values that it would actually send can
be elided -- not that any other value would not matter.

This allows servers that also host other services (like a forward proxy,
or a name-based-virtual-hosted service) to declare has-unique-proxy on
their main service.

(This was inspired by a brief idea that coap+tls and coap(s)+ws proxies
might be generally very usable as unique proxies because of SNI and Host
header -- quickly to be cruashed by the lack of protocol set in either
of these mechanisms).